Removed node dependencies from sdkModels.ts
